+++ packages/tests/lib.esm/reporter.js
@@ -71,7 +71,7 @@ export function Reporter(runner) {
     function forceOutput() {
         if (((getTime() - lastOutput) / 1000) > MAX_DELAY) {
             const currentSuite = suites[suites.length - 1];
-            log('[ Still running suite - test #' + currentSuite._countTotal + ' ]');
+            log(`[ Still running suite - test # ${(currentSuite ? currentSuite._countTotal : "0")} ]`);
         }
     }
     const timer = setInterval(forceOutput, 1000);