+++ packages/tests/lib.esm/test-contract.js
@@ -1,4 +1,13 @@
 'use strict';
+var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
+    return new (P || (P = Promise))(function (resolve, reject) {
+        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
+        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
+        step((generator = generator.apply(thisArg, _arguments || [])).next());
+    });
+};
 import assert from "assert";
 import { ethers } from "ethers";
 import contractData from "./test-contract.json";
@@ -31,7 +40,8 @@ function equals(name, actual, expected) {
     assert.equal(actual, expected, 'value matches - ' + name);
 }
 function TestContractEvents() {
-    return ethers.utils.fetchJson('https://api.ethers.io/api/v1/?action=triggerTest&address=' + contract.address).then(function (data) {
+    return __awaiter(this, void 0, void 0, function* () {
+        const data = yield ethers.utils.fetchJson('https://api.ethers.io/api/v1/?action=triggerTest&address=' + contract.address);
         console.log('  *** Triggered Transaction Hash: ' + data.hash);
         contract.on("error", (error) => {
             console.log(error);