+++ packages/tests/lib/test-contract.js
@@ -1,4 +1,40 @@
 'use strict';
+var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
+    return new (P || (P = Promise))(function (resolve, reject) {
+        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
+        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
+        step((generator = generator.apply(thisArg, _arguments || [])).next());
+    });
+};
+var __generator = (this && this.__generator) || function (thisArg, body) {
+    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
+    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
+    function verb(n) { return function (v) { return step([n, v]); }; }
+    function step(op) {
+        if (f) throw new TypeError("Generator is already executing.");
+        while (_) try {
+            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
+            if (y = 0, t) op = [op[0] & 2, t.value];
+            switch (op[0]) {
+                case 0: case 1: t = op; break;
+                case 4: _.label++; return { value: op[1], done: false };
+                case 5: _.label++; y = op[1]; op = [0]; continue;
+                case 7: op = _.ops.pop(); _.trys.pop(); continue;
+                default:
+                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
+                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
+                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
+                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
+                    if (t[2]) _.ops.pop();
+                    _.trys.pop(); continue;
+            }
+            op = body.call(thisArg, _);
+        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
+        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
+    }
+};
 var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
@@ -35,13 +71,7 @@ function equals(name, actual, expected) {
     assert_1.default.equal(actual, expected, 'value matches - ' + name);
 }
 function TestContractEvents() {
-    return ethers_1.ethers.utils.fetchJson('https://api.ethers.io/api/v1/?action=triggerTest&address=' + contract.address).then(function (data) {
-        console.log('  *** Triggered Transaction Hash: ' + data.hash);
-        contract.on("error", function (error) {
-            console.log(error);
-            assert_1.default(false);
-            contract.removeAllListeners();
-        });
+    return __awaiter(this, void 0, void 0, function () {
         function waitForEvent(eventName, expected) {
             return new Promise(function (resolve, reject) {
                 var done = false;
@@ -69,21 +99,35 @@ function TestContractEvents() {
                 }
             });
         }
-        return new Promise(function (resolve, reject) {
-            var p0 = '0x06B5955A67D827CDF91823E3bB8F069e6c89c1D6';
-            var p0_1 = '0x06b5955A67d827CdF91823e3Bb8F069e6C89C1d7';
-            var p1 = 0x42;
-            var p1_1 = 0x43;
-            return Promise.all([
-                waitForEvent('Test', [p0, p1]),
-                waitForEvent('TestP0', [p0, p1]),
-                waitForEvent('TestP0P1', [p0, p1]),
-                waitForEvent('TestIndexedString', [{ indexed: true, hash: '0x7c5ea36004851c764c44143b1dcb59679b11c9a68e5f41497f6cf3d480715331' }, p1]),
-                waitForEvent('TestV2', [{ indexed: true }, [p0, p1]]),
-                waitForEvent('TestV2Nested', [{ indexed: true }, [p0_1, p1_1, [p0, p1]]]),
-            ]).then(function (result) {
-                resolve();
-            });
+        var data;
+        return __generator(this, function (_a) {
+            switch (_a.label) {
+                case 0: return [4 /*yield*/, ethers_1.ethers.utils.fetchJson('https://api.ethers.io/api/v1/?action=triggerTest&address=' + contract.address)];
+                case 1:
+                    data = _a.sent();
+                    console.log('  *** Triggered Transaction Hash: ' + data.hash);
+                    contract.on("error", function (error) {
+                        console.log(error);
+                        assert_1.default(false);
+                        contract.removeAllListeners();
+                    });
+                    return [2 /*return*/, new Promise(function (resolve, reject) {
+                            var p0 = '0x06B5955A67D827CDF91823E3bB8F069e6c89c1D6';
+                            var p0_1 = '0x06b5955A67d827CdF91823e3Bb8F069e6C89C1d7';
+                            var p1 = 0x42;
+                            var p1_1 = 0x43;
+                            return Promise.all([
+                                waitForEvent('Test', [p0, p1]),
+                                waitForEvent('TestP0', [p0, p1]),
+                                waitForEvent('TestP0P1', [p0, p1]),
+                                waitForEvent('TestIndexedString', [{ indexed: true, hash: '0x7c5ea36004851c764c44143b1dcb59679b11c9a68e5f41497f6cf3d480715331' }, p1]),
+                                waitForEvent('TestV2', [{ indexed: true }, [p0, p1]]),
+                                waitForEvent('TestV2Nested', [{ indexed: true }, [p0_1, p1_1, [p0, p1]]]),
+                            ]).then(function (result) {
+                                resolve();
+                            });
+                        })];
+            }
         });
     });
 }