+++ packages/tests/lib/test-providers.js
@@ -560,8 +560,9 @@ function testProvider(providerName, networkName) {
         if (providerName === "NodesmithProvider") {
             return;
         }
-        // Cloudflare seems down and I want a fix out. RicMoo
-        //if (providerName === "CloudflareProvider") { return; }
+        if (providerName === "CloudflareProvider") {
+            return;
+        }
         if (networkName === "goerli" && providerName === "AlchemyProvider") {
             return;
         }